+++ misskey-util/src/pager.rs
@@ -11,6 +11,7 @@ use std::collections::VecDeque;
 use std::ops::{Deref, DerefMut};
 use std::pin::Pin;
 use std::task::{Context, Poll};
+use std::time::{Duration, Instant};
 
 use crate::Error;
 
@@ -18,9 +19,12 @@ use futures::{
     future::{BoxFuture, FutureExt},
     stream::{FusedStream, Stream, StreamExt},
 };
+use futures_timer::Delay;
 use misskey_api::{OffsetPaginationRequest, PaginationItem, PaginationRequest};
 use misskey_core::model::ApiResult;
 use misskey_core::{Client, Request};
+
+const DEFAULT_PAGE_SIZE: u8 = 30;
 
 enum PagerState<'a, C: Client + ?Sized, R: Request> {
     Pending {
@@ -42,8 +46,9 @@ impl<'a, C: Client + ?Sized, R: PaginationRequest> BackwardPager<'a, C, R> {
     pub(crate) fn with_since_id(
         client: &'a C,
         since_id: Option<<R::Item as PaginationItem>::Id>,
-        request: R,
+        mut request: R,
     ) -> Self {
+        request.set_limit(DEFAULT_PAGE_SIZE);
         BackwardPager {
             client,
             since_id,
@@ -54,13 +59,7 @@ impl<'a, C: Client + ?Sized, R: PaginationRequest> BackwardPager<'a, C, R> {
     }
 
     pub(crate) fn new(client: &'a C, request: R) -> Self {
-        BackwardPager {
-            client,
-            since_id: None,
-            state: Some(PagerState::Ready {
-                next_request: request,
-            }),
-        }
+        BackwardPager::with_since_id(client, None, request)
     }
 }
 
@@ -137,6 +136,14 @@ where
 {
     type Content = R::Item;
     type Client = C;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        match self.state.as_mut() {
+            Some(PagerState::Ready { next_request, .. }) => next_request.set_limit(size),
+            Some(PagerState::Pending { request, .. }) => request.set_limit(size),
+            None => {}
+        }
+    }
 }
 
 pub(crate) struct ForwardPager<'a, C: Client + ?Sized, R: Request> {
@@ -145,7 +152,8 @@ pub(crate) struct ForwardPager<'a, C: Client + ?Sized, R: Request> {
 }
 
 impl<'a, C: Client + ?Sized, R: PaginationRequest> ForwardPager<'a, C, R> {
-    pub(crate) fn new(client: &'a C, request: R) -> Self {
+    pub(crate) fn new(client: &'a C, mut request: R) -> Self {
+        request.set_limit(DEFAULT_PAGE_SIZE);
         ForwardPager {
             client,
             state: Some(PagerState::Ready {
@@ -221,6 +229,14 @@ where
 {
     type Content = R::Item;
     type Client = C;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        match self.state.as_mut() {
+            Some(PagerState::Ready { next_request, .. }) => next_request.set_limit(size),
+            Some(PagerState::Pending { request, .. }) => request.set_limit(size),
+            None => {}
+        }
+    }
 }
 
 pub(crate) struct OffsetPager<'a, C: Client + ?Sized, R: Request> {
@@ -230,7 +246,8 @@ pub(crate) struct OffsetPager<'a, C: Client + ?Sized, R: Request> {
 }
 
 impl<'a, C: Client + ?Sized, R: OffsetPaginationRequest> OffsetPager<'a, C, R> {
-    pub(crate) fn new(client: &'a C, request: R) -> Self {
+    pub(crate) fn new(client: &'a C, mut request: R) -> Self {
+        request.set_limit(DEFAULT_PAGE_SIZE);
         OffsetPager {
             client,
             state: Some(PagerState::Ready {
@@ -308,6 +325,14 @@ where
 {
     type Content = R::Item;
     type Client = C;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        match self.state.as_mut() {
+            Some(PagerState::Ready { next_request, .. }) => next_request.set_limit(size),
+            Some(PagerState::Pending { request, .. }) => request.set_limit(size),
+            None => {}
+        }
+    }
 }
 
 /// A stream of pages..
@@ -320,16 +345,27 @@ pub trait Pager:
     type Content;
     /// [`Client`][`misskey_core::Client`] type used in the pager.
     type Client: Client + ?Sized;
+
+    /// Sets the number of items to be fetched at once.
+    fn set_page_size(self: Pin<&mut Self>, size: u8);
 }
 
 impl<P: Pager + Unpin + ?Sized> Pager for &mut P {
     type Content = P::Content;
     type Client = P::Client;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        P::set_page_size(Pin::new(&mut **self), size)
+    }
 }
 
 impl<P: Pager + Unpin + ?Sized> Pager for Box<P> {
     type Content = P::Content;
     type Client = P::Client;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        P::set_page_size(Pin::new(&mut **self), size)
+    }
 }
 
 impl<P> Pager for Pin<P>
@@ -339,15 +375,23 @@ where
 {
     type Content = <<P as Deref>::Target as Pager>::Content;
     type Client = <<P as Deref>::Target as Pager>::Client;
+
+    fn set_page_size(self: Pin<&mut Self>, size: u8) {
+        <<P as Deref>::Target as Pager>::set_page_size(self.get_mut().as_mut(), size)
+    }
 }
 
 impl<S, F, T> Pager for futures::stream::MapOk<S, F>
 where
-    S: Pager,
+    S: Pager + Unpin,
     F: FnMut(Vec<<S as Pager>::Content>) -> Vec<T>,
 {
     type Content = T;
     type Client = <S as Pager>::Client;
+
+    fn set_page_size(mut self: Pin<&mut Self>, size: u8) {
+        <S as Pager>::set_page_size(Pin::new(&mut *(*self).get_mut()), size)
+    }
 }
 
 /// An owned dynamically typed [`Pager`].
@@ -359,18 +403,46 @@ pub type BoxPager<'a, C, T> = Pin<
     >,
 >;
 
+enum PagerStreamState<P: Pager> {
+    Ready {
+        item: P::Content,
+        buffer: VecDeque<P::Content>,
+        last_fetch: Instant,
+    },
+    Fetch,
+    Wait(Delay),
+}
+
 /// A stream of elements in [`Pager`].
 pub struct PagerStream<P: Pager> {
     pager: P,
-    buffer: VecDeque<P::Content>,
+    minimum_interval: Duration,
+    state: Option<PagerStreamState<P>>,
 }
 
 impl<P: Pager> PagerStream<P> {
     pub(crate) fn new(pager: P) -> Self {
         PagerStream {
             pager,
-            buffer: VecDeque::new(),
+            minimum_interval: Duration::new(0, 0),
+            state: Some(PagerStreamState::Fetch),
         }
+    }
+
+    /// Sets the number of items to be fetched at once by the inner pager.
+    pub fn set_page_size(&mut self, size: u8)
+    where
+        P: Unpin,
+    {
+        Pin::new(&mut self.pager).set_page_size(size);
+    }
+
+    /// Sets the minimum time interval between pagination.
+    ///
+    /// It is recommended to set this to reduce the load on the
+    /// server if you expect a lot of pagination.
+    pub fn set_interval(&mut self, minimum_interval: Duration) {
+        self.minimum_interval = minimum_interval;
     }
 
     /// Returns the inner pager.
@@ -387,13 +459,61 @@ where
     type Item = Result<P::Content, Error<<P::Client as Client>::Error>>;
 
     fn poll_next(mut self: Pin<&mut Self>, cx: &mut Context) -> Poll<Option<Self::Item>> {
-        if let Some(item) = self.buffer.pop_front() {
-            Poll::Ready(Some(Ok(item)))
-        } else if let Some(page) = futures::ready!(self.pager.poll_next_unpin(cx)) {
-            self.buffer.extend(page?);
-            Poll::Ready(self.buffer.pop_front().map(Ok))
-        } else {
-            Poll::Ready(None)
+        let state = self.state.take();
+        match state {
+            Some(PagerStreamState::Ready {
+                item,
+                mut buffer,
+                last_fetch,
+            }) => {
+                if let Some(next) = buffer.pop_front() {
+                    self.state = Some(PagerStreamState::Ready {
+                        item: next,
+                        buffer,
+                        last_fetch,
+                    });
+                } else {
+                    let until = last_fetch + self.minimum_interval;
+                    if let Some(duration) = until.checked_duration_since(Instant::now()) {
+                        self.state = Some(PagerStreamState::Wait(Delay::new(duration)));
+                    } else {
+                        self.state = Some(PagerStreamState::Fetch);
+                    }
+                }
+                Poll::Ready(Some(Ok(item)))
+            }
+            Some(PagerStreamState::Fetch) => match self.pager.poll_next_unpin(cx) {
+                Poll::Pending => {
+                    self.state = Some(PagerStreamState::Fetch);
+                    Poll::Pending
+                }
+                Poll::Ready(None) => Poll::Ready(None),
+                Poll::Ready(Some(Err(err))) => Poll::Ready(Some(Err(err))),
+                Poll::Ready(Some(Ok(page))) => {
+                    let mut buffer: VecDeque<_> = page.into();
+                    if let Some(item) = buffer.pop_front() {
+                        self.state = Some(PagerStreamState::Ready {
+                            item,
+                            buffer,
+                            last_fetch: Instant::now(),
+                        });
+                        self.poll_next(cx)
+                    } else {
+                        Poll::Ready(None)
+                    }
+                }
+            },
+            Some(PagerStreamState::Wait(mut delay)) => match delay.poll_unpin(cx) {
+                Poll::Pending => {
+                    self.state = Some(PagerStreamState::Wait(delay));
+                    Poll::Pending
+                }
+                Poll::Ready(()) => {
+                    self.state = Some(PagerStreamState::Fetch);
+                    self.poll_next(cx)
+                }
+            },
+            None => Poll::Ready(None),
         }
     }
 }
@@ -404,6 +524,6 @@ where
     P::Content: Unpin + std::fmt::Debug,
 {
     fn is_terminated(&self) -> bool {
-        self.buffer.is_empty() && self.pager.is_terminated()
+        self.state.is_none()
     }
 }